Refactored simple-httpd, introduce CLI and simpler interface
This gem rebuilds Simple::Httpd to introduce a cleaner interface. Now the public interface is

    Simple::Httpd.build(mounts)

which accepts an array of mountpoint descriptions, and returns a rack app, which mixes handlers to both dynamic and static content.

With this change simple-httpd also introduces a simple-httpd gem, which wraps the functionality into an easy to use script.

This commit also adds documentation and specs