Implements “simple-http routes” command
When building the simple-httpd application we also collect all defined routes, together with their source location (which, for example, might point to a service’s method or to a controller source file).

This information is returned by the `route_descriptions` method.

The “simple-http routes” subcommand prints all routes of a constructed simple-http application.